pyproject.toml
[tool.poetry]
name = "magic-dashboard"
version = "0.1.0"
description = "dashboard and tracker"
authors = ["mysubsnewslivecom <my.subs.news@live.com>"]
readme = "README.md"
# packages = [{include = "magic_dashboard"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
psycopg2 = "^2.9.5"
dj-database-url = "^1.2.0"
loguru = "^0.6.0"
requests = "^2.28.1"
pillow = "^9.3.0"
django-redis = "^5.2.0"
python-dotenv = "^0.21.0"
django-crispy-forms = "^1.14.0"
djangorestframework = ">=3.13.1,<3.14.0 || >3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular-sidecar = "^2022.12.1"
drf-spectacular = "^0.25.1"
lxml = "^4.9.2"
beautifulsoup4 = "^4.11.1"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
black = "^22.12.0"
isort = "^5.11.4"
yapf = "^0.32.0"
flake8 = "^6.0.0"
autopep8 = "^2.0.1"
autoflake = "^2.0.0"
django-tables2 = "^2.5.0"
python-gitlab = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
